roundtable/requirements.txt
# RoundTable CLI 依赖
# MVP 最小依赖集

# HTTP 客户端（模型调用）
httpx>=0.26.0

# FastAPI（未来 Web UI）
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9

# Qdrant 向量数据库
qdrant-client>=1.7.0

# 文件处理
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# 模型接入
openai>=1.10.0        # OpenRouter 兼容（GPT/DeepSeek）
google-genai>=0.3.0   # Gemini 直连

# 工具依赖
python-dotenv>=1.0.0  # .env 文件加载
pydantic>=2.5.0       # 数据验证

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Embedding 模型（BGE-M3 本地向量化）
FlagEmbedding>=1.2.0  # BGE 官方库（首选）
sentence-transformers>=2.3.0  # 备选降级方案
